requirements.txt
# ── Interface gráfica ──────────────────────────────────────────────────────
customtkinter>=5.2.2
Pillow>=10.0.0

# ── Google Drive API ───────────────────────────────────────────────────────
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
googleapis-common-protos>=1.62.0

# ── Monitoramento de arquivos ──────────────────────────────────────────────
watchdog>=4.0.0

# ── Criptografia de credenciais ────────────────────────────────────────────
cryptography>=41.0.0

# ── Bandeja do sistema (Linux) ─────────────────────────────────────────────
pystray>=0.19.5

# ── Notificações desktop (opcional — fallback para notify-send) ────────────
plyer>=2.1.0

# ── Utilitários ───────────────────────────────────────────────────────────
requests>=2.31.0
